requirements.txt
aiohttp
aiosignal
APScheduler
async-timeout
asyncio
attrs
aws-lambda-powertools
aws-xray-sdk
beautifulsoup4
blinker
boto3
botocore
certifi
charset-normalizer
click
contourpy
cycler
DateTime
easyocr
exceptiongroup
fastjsonschema
filelock
Flask
Flask-Ext
Flask-Mail
Flask-Script
Flask-SQLAlchemy
fonttools
frozenlist
fsspec
fyers-api
fyers-apiv3
greenlet
gunicorn
idna
imageio
importlib_metadata
importlib_resources
iniconfig
itsdangerous
Jinja2
jmespath
joblib
json5
jsonschema
jsonschema-specifications
kiwisolver
lazy_loader
MarkupSafe
matplotlib
mplfinance
mpmath
multidict
mysql-connector-python
nbformat
networkx
ninja
numpy
opencv-python
opencv-python-headless
packaging
pandas
pandas_ta
pillow
plotly
pluggy
pyclipper
pyotp
pyparsing
pytesseract
pytest
pytest-warnings
python-bidi
pytz
PyYAML
referencing
requests
rpds-py
s3transfer
scikit-image
scikit-learn
scipy
shapely
six
soupsieve
SQLAlchemy
sympy
tenacity
threadpoolctl
tifffile
tomli
torch
torchvision
typing_extensions
tzdata
tzlocal
urllib3
websocket-client
Werkzeug
wrapt
yarl
zipp
zope.interface